test: move into single file
