+++ packages/nuxt-cli/test/e2e/commands.spec.ts
@@ -160,3 +160,25 @@ describe('commands', () => {
     }
   }
 })
+
+describe('extends support', () => {
+  it('works with dev server', { timeout: isWindows ? 200000 : 50000 }, async () => {
+    const controller = new AbortController()
+    const port = await getPort({ host: '127.0.0.1', port: 3003 })
+    const devProcess = x(nuxi, ['dev', `--host=127.0.0.1`, `--port=${port}`, '--extends=some-layer'], {
+      nodeOptions: { stdio: 'pipe', cwd: fixtureDir },
+      signal: controller.signal,
+    })
+
+    // Test that server responds
+    const response = await fetchWithPolling(`http://127.0.0.1:${port}/extended`, {}, 30, 300)
+    expect.soft(response?.status).toBe(200)
+    expect(await response?.text()).toContain('This is an extended page from a layer.')
+
+    controller.abort()
+    try {
+      await devProcess
+    }
+    catch {}
+  })
+})
